% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_distinct.R
\name{count_distinct}
\alias{count_distinct}
\title{Count distinct values}
\usage{
count_distinct(x, ..., distinct_by, sort = FALSE, name = "n", na.rm = FALSE)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{...}{Variables to group by.}

\item{distinct_by}{The name of the column to count distinct values of.}

\item{sort}{If TRUE, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.}

\item{na.rm}{If TRUE, exclude missing observations from the count.
If there are multiple vectors in ..., an observation will be excluded if any
of the values are missing.}
}
\description{
Count distinct values
}
