% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_df.R
\name{search_tbl}
\alias{search_tbl}
\title{Search a data frame for a value}
\usage{
search_tbl(x, value)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{value}{The value to search for.}
}
\description{
Search a data frame for a value
}
